+++ src/components/ShareForm/DateContainer/DateContainer.style.tsx
@@ -3,8 +3,6 @@ import 'react-datepicker/dist/react-datepicker.css';
 import ReactDatePicker from 'react-datepicker';
 import styled, { css } from 'styled-components';
 
-import { flexBetween, flexCenter, noScrollBar } from '@styles/mixin';
-
 const DatePickerStyle = css`
   ${({ theme: { fonts, colors } }) => css`
     position: relative;
@@ -179,109 +177,6 @@ const TimePickerStyle = css`
   `}
 `;
 
-//FileContainer.tsx
-export const FileWrapper = styled.div`
-  display: flex;
-  gap: 12px;
-  width: calc(100% + 1rem);
-  padding-top: 15px;
-  padding-right: 1rem;
-  overflow-x: scroll;
-  scroll-behavior: smooth;
-  ${noScrollBar}
-`;
-
-export const FileLabel = styled.label<{ isImage: boolean }>`
-  ${({ theme: { colors, fonts }, isImage }) => css`
-    ${fonts.smallRegular};
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 0.25rem;
-    cursor: pointer;
-    width: 4.5rem;
-    min-width: 4.5rem;
-    height: 4.5rem;
-    background-color: ${colors.white2};
-    border: 1px solid ${colors.grey2};
-    border-radius: 4px;
-    path {
-      fill: ${isImage && colors.orange2};
-    }
-  `}
-`;
-
-export const FileCount = styled.div`
-  color: ${({ theme }) => theme.colors.grey4};
-`;
-
-export const FileLength = styled.span<{ isFile?: number }>`
-  ${({ isFile }) => css`
-    ${isFile &&
-    css`
-      color: ${({ theme }) => theme.colors.orange3};
-    `}
-  `}
-`;
-
-export const FileForm = styled.input`
-  display: none;
-`;
-
-export const ImagePreviewContainer = styled.div`
-  position: relative;
-
-  svg {
-    position: absolute;
-    top: -4px;
-    right: -4px;
-  }
-`;
-
-//default style
-export const TwoTextBlock = styled.div`
-  display: flex;
-  gap: 12px;
-  width: 100%;
-  justify-content: flex-start;
-`;
-
-export const LongTextBlock = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-`;
-
-//TextContainer.tsx
-export const TextContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-`;
-
-//AddressContainer.tsx
-export const LocationSelectButton = styled.button`
-  ${flexBetween}
-  padding: 0 12px;
-  width: 100%;
-  height: 48px;
-  border: 1px solid ${({ theme }) => theme.colors.grey3};
-  border-radius: 4px;
-  svg {
-    padding-right: 0.2rem;
-  }
-`;
-
-export const NoneLocation = styled.span`
-  ${({ theme: { fonts, colors } }) => css`
-    ${fonts.largeRegular};
-
-    color: ${colors.grey4};
-  `}
-`;
-
-//DateContainer.tsx
 export const DateInputWrapper = styled.label<{ isDateFocused: boolean }>`
   ${({ theme: { colors }, isDateFocused }) => css`
     color: ${colors.black};
@@ -412,79 +307,6 @@ export const TimeInputForm = styled.input`
     ::-webkit-calendar-picker-indicator {
       display: none;
       -webkit-appearance: none;
-    }
-  `}
-`;
-
-//RecruitmentContainer.tsx
-export const CountContainer = styled.div`
-  display: flex;
-  align-items: center;
-  width: 60%;
-  gap: 10px;
-`;
-
-export const ButtonContainer = styled.div`
-  ${({ theme: { colors, fonts } }) => css`
-    ${flexCenter}
-    ${fonts.large}
-
-    height: 48px;
-    width: 96px;
-    padding: 2px 4px;
-    display: flex;
-    justify-content: space-around;
-    border: 1px solid ${colors.grey3};
-    border-radius: 4px;
-  `}
-`;
-
-export const CountButton = styled.button<{ isClickable: boolean }>`
-  ${({ theme: { colors }, isClickable }) => css`
-    width: 24px;
-    height: 24px;
-    path {
-      stroke: ${isClickable ? colors.black : colors.grey4};
-    }
-  `}
-`;
-
-export const RecruitmentText = styled.span`
-  ${({ theme: { colors, fonts } }) => css`
-    ${fonts.largeRegular};
-    color: ${colors.grey4};
-  `}
-`;
-
-//ContentDescription.tsx
-export const DescriptionWrapper = styled.div`
-  min-height: 236px;
-`;
-
-export const ContentDescriptionInput = styled.textarea`
-  ${({ theme: { colors, fonts } }) => css`
-    ${fonts.main};
-    ${fonts.largeRegular};
-    width: 100%;
-    min-height: 236px;
-    padding: 0.75rem;
-    border: 1px solid ${colors.grey3};
-    background: ${colors.white2};
-    border-radius: 4px;
-    resize: none;
-
-    :focus-visible {
-      outline: none;
-    }
-
-    :focus {
-      ::placeholder {
-        color: transparent;
-      }
-    }
-
-    ::placeholder {
-      color: ${colors.grey4};
     }
   `}
 `;
