+++ src/components/ShareForm/ShareForm.style.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const TwoTextBlock = styled.div`
+  display: flex;
+  gap: 12px;
+  width: 100%;
+  justify-content: flex-start;
+`;
+
+export const LongTextBlock = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+`;
