+++ src/components/ShareForm/TextContainer/TextContainer.style.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const TextContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+`;
